snaplink/requirements.txt
Django==5.2.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
mysqlclient==2.2.8
redis==7.2.0
django-redis==6.0.0
django-ratelimit==4.1.0
python-dotenv==1.0.0
user-agents==2.2.0
gunicorn==21.2.0